Extract MR-JAR manifest processing into standalone class
